fixe the bug of reading file in. It used to read the decimal in and made it a integer. However, this should not have happen. I fixed it by changing the method used from atoi() to atof()

+++ p4/user_io.cpp
@@ -60,9 +60,9 @@ void readCurves(std::ifstream *ifs, Graph *graph, std::list<Curve*> *curves, int
       for(int i = 0 ; i < numberOfCtrlPoints; i++){
         ifs->getline(buffer, sizeOfBuffer);
         charPtr = strtok(buffer, ", ");
-        x = atoi(charPtr); 
+        x = atof(charPtr); 
         charPtr = strtok(0, "\0");
-        y = atoi(charPtr);
+        y = atof(charPtr);
         ctrlPoints[i].x = x;
         ctrlPoints[i].y = y;
       }
@@ -85,9 +85,9 @@ void readCurves(std::ifstream *ifs, Graph *graph, std::list<Curve*> *curves, int
       for(int i = 0 ; i < numberOfCtrlPoints; i++){
         ifs->getline(buffer, sizeOfBuffer);
         charPtr = strtok(buffer, ", ");
-        x = atoi(charPtr); 
+        x = atof(charPtr); 
         charPtr = strtok(0, "\0");
-        y = atoi(charPtr);
+        y = atof(charPtr);
         ctrlPoints[i].x = x;
         ctrlPoints[i].y = y;
       }
